some changes

+++ src/Algorithms/AbstractAlgorithm.h
@@ -2,7 +2,6 @@
 #define A3_PARALLELS_0_MASTER_ABSTRACT_ALGORITHM_H
 
 #include <vector>
-#include "time.h"
 
 #include "../DataStructures/Matrix/Matrix.h"
 
@@ -12,7 +11,7 @@ public:
     virtual ~AbstractAlgorithm() {}
     virtual S21Matrix SolveWithoutUsingParallelism(std::vector<S21Matrix> matrices) = 0;
     virtual S21Matrix SolveUsingParallelism(std::vector<S21Matrix> matrices) = 0;
-    std::pair<int, int> MeasureTime(
+    std::pair<double, double> MeasureTime(
         std::vector<S21Matrix> matrices,
         int number_of_repetitions);  //  возвращает время решения без параллелизма и с
 };
